Update Distribute To Close Match Spec
First pass at updating the distribute code to match the spec closer.  It is
not in full compliance because there are some required that are not mentioned
in the spec. 'filter_description' and 'resource_data_url' are additional
fields that are source node needs for the destination to complete the replication